fix(dashboard): address code review findings
- Fix f64→i64 truncation in ThresholdConfig (data loss for noop values)
- Replace silent unwrap_or_default with expect in YAML serializer
- Add assertions to Grid::grid() preventing mixed entries and zero cols

+++ src/lib/dashboard/widgets.rs
@@ -173,7 +173,7 @@ impl ThresholdConfig {
         let mut m = Mapping::new();
         m.insert(
             Value::String("value".into()),
-            Value::Number(serde_yaml::Number::from(self.value as i64)),
+            Value::Number(serde_yaml::Number::from(self.value)),
         );
         m.insert(
             Value::String("color".into()),
